requirements.txt
streamlit
pandas
python-docx
openpyxl